fix: update settings type to be an aray
