+++ src/module/types.d.ts
@@ -293,3 +293,11 @@ declare global {
 		);
 	}
 }
+export type ShareTargetSettingsOptions = {
+	send: boolean;
+	recieve: boolean;
+	index: number;
+	name: string;
+	role: string;
+	color: string;
+} & Partial<FormApplicationOptions>;
