--- src/types/settings.ts
@@ -1,8 +1,0 @@
-export type ShareTargetSettingsOptions = {
-	send: boolean;
-	recieve: boolean;
-	index: number;
-	name: string;
-	role: string;
-	color: string;
-} & Partial<FormApplicationOptions>;
